Modified operator to handle clusters in multiple namespace
